refactor(components): moved IconButton to own file
