+++ app/components/index.ts
@@ -10,3 +10,5 @@ export * from "./TextField"
 export * from "./Toggle"
 export * from "./EmptyState"
 export * from "./Avatar"
+export * from "./Tag"
+export * from "./IconButton"
